CHANGED
I removed the background from the draw function so that it fades away slowly but I altered the code so that the background is reset whenever the ball hits the paddle. I have also made it so that the user can move the paddle anywhere on the screen to hit the ball.