All of the changes all at once
While working on this project, I had forgotten to push my changes and it was too difficult to try to redo the process.
I have added static and linear patterns to the background, changed the ball's movement to create a wavelength type effect as well as making it change from white to black as it moves from one half of the screen to the next. I have also made two scores, one depicting how many points the player has accumulated and how many times the player has missed the ball. Once the player misses the ball 10 times, the game will end, showing the game over screen with the number of points the player has accumulated. I have changed this two player game to a one-player game where the player uses the numbers 1 and 0 to move the paddles up and down the paddles move in the opposite direction from each other as they move up and down the screen. the point of this game is for the player to accumulate as many points as possible before he/she misses the ball a total of 10 times.