Created a bouncer + matched sound file to bouncer
I have created a new ball that bounces around twithin the screen. Whenever the ball bounces against the top or bottom of the window, the sound will be commanded. This will serve as a base tempo. I am still trying to change the velocity with my up and down keys to allow the user to make the tempo faster or slower.